Moved function folder into common folder
