+++ src/common/functions/checkIsValidImageUrl.js
